feat: add config for keep email and top folder changes

+++ src/file_mover_for_google_drive/actions/apply.py
@@ -315,6 +315,18 @@ class Apply(manage.BaseManage):
                 }
             )
 
+        if (
+            plan.begin_user_email in actions_config.permissions_user_emails_keep
+            or plan.end_user_email in actions_config.permissions_user_emails_keep
+        ):
+            return report.OutcomeReport(
+                **{
+                    **params,
+                    "result_name": models.PlanReportOutcomes.SKIPPED,
+                    "result_description": "Config prevented executing plan action",
+                }
+            )
+
         # execute the plan
         actions.delete_permission(plan.entry_id, plan.permission_id)
